refactor: mcp server form
