+++ renderer/src/features/mcp-servers/components/card-mcp-server/server-actions/items/edit-configuration-menu-item.tsx
@@ -1,7 +1,7 @@
 import { Settings } from 'lucide-react'
 import { useState } from 'react'
 import { DropdownMenuItem } from '@/common/components/ui/dropdown-menu'
-import { DeprecatedDialogFormRunMcpServerWithCommand } from '../../../dialog-form-run-mcp-command'
+import { DialogFormLocalMcp } from '../../../local-mcp/dialog-form-local-mcp'
 
 interface EditConfigurationMenuItemProps {
   serverName: string
@@ -26,7 +26,7 @@ export function EditConfigurationMenuItem({
         </a>
       </DropdownMenuItem>
 
-      <DeprecatedDialogFormRunMcpServerWithCommand
+      <DialogFormLocalMcp
         isOpen={isRunWithCommandOpen}
         onOpenChange={setIsRunWithCommandOpen}
         serverToEdit={serverName}
