+++ renderer/src/features/mcp-servers/components/local-mcp/__tests__/dialog-form-local-mcp.test.tsx
@@ -1,16 +1,16 @@
 import { render, waitFor, screen, act } from '@testing-library/react'
 import { it, expect, vi, describe, beforeEach } from 'vitest'
-import { DeprecatedDialogFormRunMcpServerWithCommand } from '../dialog-form-run-mcp-command'
+import { DialogFormLocalMcp } from '../dialog-form-local-mcp'
 import userEvent from '@testing-library/user-event'
 import { Dialog } from '@/common/components/ui/dialog'
 import { server as mswServer } from '@/common/mocks/node'
 import { http, HttpResponse } from 'msw'
 import { mswEndpoint } from '@/common/mocks/msw-endpoint'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import { useRunCustomServer } from '../../hooks/use-run-custom-server'
+import { useRunCustomServer } from '../../../hooks/use-run-custom-server'
 
 // Mock the hook
-vi.mock('../../hooks/use-run-custom-server', () => ({
+vi.mock('../../../hooks/use-run-custom-server', () => ({
   useRunCustomServer: vi.fn(),
 }))
 
@@ -67,14 +67,11 @@ beforeEach(() => {
   })
 })
 
-describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
+describe('DialogFormLocalMcp', () => {
   it('renders form fields correctly for docker image', async () => {
     renderWithProviders(
       <Wrapper>
-        <DeprecatedDialogFormRunMcpServerWithCommand
-          isOpen
-          onOpenChange={vi.fn()}
-        />
+        <DialogFormLocalMcp isOpen closeDialog={vi.fn()} />
       </Wrapper>
     )
 
@@ -116,10 +113,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
 
     renderWithProviders(
       <Wrapper>
-        <DeprecatedDialogFormRunMcpServerWithCommand
-          isOpen
-          onOpenChange={vi.fn()}
-        />
+        <DialogFormLocalMcp isOpen closeDialog={vi.fn()} />
       </Wrapper>
     )
 
@@ -173,10 +167,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
 
     renderWithProviders(
       <Wrapper>
-        <DeprecatedDialogFormRunMcpServerWithCommand
-          isOpen
-          onOpenChange={vi.fn()}
-        />
+        <DialogFormLocalMcp isOpen closeDialog={vi.fn()} />
       </Wrapper>
     )
 
@@ -232,10 +223,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
 
     renderWithProviders(
       <Wrapper>
-        <DeprecatedDialogFormRunMcpServerWithCommand
-          isOpen
-          onOpenChange={vi.fn()}
-        />
+        <DialogFormLocalMcp isOpen closeDialog={vi.fn()} />
       </Wrapper>
     )
 
@@ -319,10 +307,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
 
     renderWithProviders(
       <Wrapper>
-        <DeprecatedDialogFormRunMcpServerWithCommand
-          isOpen
-          onOpenChange={vi.fn()}
-        />
+        <DialogFormLocalMcp isOpen closeDialog={vi.fn()} />
       </Wrapper>
     )
 
@@ -362,10 +347,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
 
     renderWithProviders(
       <Wrapper>
-        <DeprecatedDialogFormRunMcpServerWithCommand
-          isOpen
-          onOpenChange={vi.fn()}
-        />
+        <DialogFormLocalMcp isOpen closeDialog={vi.fn()} />
       </Wrapper>
     )
 
@@ -397,10 +379,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
   it('renders aria-hidden column labels for storage volumes', async () => {
     renderWithProviders(
       <Wrapper>
-        <DeprecatedDialogFormRunMcpServerWithCommand
-          isOpen
-          onOpenChange={vi.fn()}
-        />
+        <DialogFormLocalMcp isOpen closeDialog={vi.fn()} />
       </Wrapper>
     )
 
@@ -433,10 +412,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
 
     renderWithProviders(
       <Wrapper>
-        <DeprecatedDialogFormRunMcpServerWithCommand
-          isOpen
-          onOpenChange={mockOnOpenChange}
-        />
+        <DialogFormLocalMcp isOpen closeDialog={mockOnOpenChange} />
       </Wrapper>
     )
 
@@ -471,7 +447,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
 
     await waitFor(() => {
       expect(mockCheckServerStatus).toHaveBeenCalled()
-      expect(mockOnOpenChange).toHaveBeenCalledWith(false)
+      expect(mockOnOpenChange).toHaveBeenCalled()
     })
   })
 
@@ -480,10 +456,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
 
     renderWithProviders(
       <Wrapper>
-        <DeprecatedDialogFormRunMcpServerWithCommand
-          isOpen
-          onOpenChange={mockOnOpenChange}
-        />
+        <DialogFormLocalMcp isOpen closeDialog={mockOnOpenChange} />
       </Wrapper>
     )
 
@@ -493,7 +466,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
 
     await userEvent.click(screen.getByRole('button', { name: 'Cancel' }))
 
-    expect(mockOnOpenChange).toHaveBeenCalledWith(false)
+    expect(mockOnOpenChange).toHaveBeenCalled()
   })
 
   describe('Network Isolation', () => {
@@ -508,10 +481,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
 
       renderWithProviders(
         <Wrapper>
-          <DeprecatedDialogFormRunMcpServerWithCommand
-            isOpen
-            onOpenChange={vi.fn()}
-          />
+          <DialogFormLocalMcp isOpen closeDialog={vi.fn()} />
         </Wrapper>
       )
 
@@ -576,10 +546,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
 
       renderWithProviders(
         <Wrapper>
-          <DeprecatedDialogFormRunMcpServerWithCommand
-            isOpen
-            onOpenChange={vi.fn()}
-          />
+          <DialogFormLocalMcp isOpen closeDialog={vi.fn()} />
         </Wrapper>
       )
 
@@ -618,10 +585,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
 
       renderWithProviders(
         <Wrapper>
-          <DeprecatedDialogFormRunMcpServerWithCommand
-            isOpen
-            onOpenChange={mockOnOpenChange}
-          />
+          <DialogFormLocalMcp isOpen closeDialog={mockOnOpenChange} />
         </Wrapper>
       )
 
@@ -635,14 +599,11 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
 
       await userEvent.click(screen.getByRole('button', { name: 'Cancel' }))
 
-      expect(mockOnOpenChange).toHaveBeenCalledWith(false)
+      expect(mockOnOpenChange).toHaveBeenCalled()
 
       renderWithProviders(
         <Wrapper>
-          <DeprecatedDialogFormRunMcpServerWithCommand
-            isOpen
-            onOpenChange={mockOnOpenChange}
-          />
+          <DialogFormLocalMcp isOpen closeDialog={mockOnOpenChange} />
         </Wrapper>
       )
 
@@ -657,10 +618,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
     it('activates the network isolation tab if a validation error occurs there while on the configuration tab', async () => {
       renderWithProviders(
         <Wrapper>
-          <DeprecatedDialogFormRunMcpServerWithCommand
-            isOpen
-            onOpenChange={vi.fn()}
-          />
+          <DialogFormLocalMcp isOpen closeDialog={vi.fn()} />
         </Wrapper>
       )
 
@@ -707,10 +665,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
     it('activates the configuration tab if a validation error occurs there while on the network isolation tab', async () => {
       renderWithProviders(
         <Wrapper>
-          <DeprecatedDialogFormRunMcpServerWithCommand
-            isOpen
-            onOpenChange={vi.fn()}
-          />
+          <DialogFormLocalMcp isOpen closeDialog={vi.fn()} />
         </Wrapper>
       )
 
@@ -733,10 +688,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
     it('shows alert when network isolation is enabled but no hosts or ports are configured', async () => {
       renderWithProviders(
         <Wrapper>
-          <DeprecatedDialogFormRunMcpServerWithCommand
-            isOpen
-            onOpenChange={vi.fn()}
-          />
+          <DialogFormLocalMcp isOpen closeDialog={vi.fn()} />
         </Wrapper>
       )
 
@@ -785,10 +737,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
 
       renderWithProviders(
         <Wrapper>
-          <DeprecatedDialogFormRunMcpServerWithCommand
-            isOpen
-            onOpenChange={vi.fn()}
-          />
+          <DialogFormLocalMcp isOpen closeDialog={vi.fn()} />
         </Wrapper>
       )
 
@@ -879,10 +828,7 @@ describe('DeprecatedDialogFormRunMcpServerWithCommand', () => {
 
     renderWithProviders(
       <Wrapper>
-        <DeprecatedDialogFormRunMcpServerWithCommand
-          isOpen
-          onOpenChange={vi.fn()}
-        />
+        <DialogFormLocalMcp isOpen closeDialog={vi.fn()} />
       </Wrapper>
     )
